+++ ui/src/components/search-form.tsx
@@ -7,49 +7,14 @@ import {
   ArrowsRightLeftIcon,
 } from '@heroicons/react/24/outline';
 
-import { ChangeEventHandler, ReactNode } from 'react';
-import clsx from 'clsx';
+import { ChangeEventHandler } from 'react';
 
 import Text from '@/components/atomic/text';
 import { useSearchFormState } from '@/hooks/use-search-form-state';
 import { useDropdown } from '@/hooks/use-dropdown';
 import { FiltersProvider, useFilters } from '@/context/filters-context';
 import { useSearchParams } from 'next/navigation';
-
-// TODO: make it generic and seperate
-function ChipButton({
-  onClick,
-  Text,
-  Icon,
-  rounded,
-  type = 'button',
-  width,
-}: {
-  Text: ReactNode;
-  Icon?: ReactNode;
-  onClick?: () => void;
-  rounded: 'full' | 'lg';
-  type?: 'button' | 'submit';
-  width: 'fit' | 'full';
-}) {
-  const roundClass = clsx({ 'rounded-lg': rounded === 'lg', 'rounded-full': rounded === 'full' });
-  const widthClass = clsx({ 'w-full': width === 'full', 'w-fit': width === 'fit' });
-
-  return (
-    <button
-      className={clsx(
-        'flex items-center space-x-1 px-2 py-1 hover:bg-gray-100',
-        roundClass,
-        widthClass
-      )}
-      onClick={onClick}
-      type={type}
-    >
-      {Icon}
-      {Text}
-    </button>
-  );
-}
+import ChipButton from '@/components/atomic/chip-button';
 
 function FilterInput({ filter }: { filter: string }) {
   const {
